+++ include/num_collect/interp/kernel/kernel_interpolator.h
@@ -21,13 +21,17 @@
 
 #include <vector>
 
-#include "num_collect/interp/kernel/impl/kernel_interpolator_impl.h"
-#include "num_collect/interp/kernel/impl/kernel_parameter_optimizer.h"
+#include "num_collect/interp/kernel/impl/kernel_coeff_solver.h"
 
 namespace num_collect::interp::kernel {
 
 /*!
  * \brief Class to interpolate data using kernels.
+ *
+ * This class determines parameters automatically if set to do so
+ * using maximum likelihood estimation (MLE) \cite Scheuerer2011.
+ * And this class can evaluate mean and variance in Gaussian process as in
+ * \cite Brochu2010.
  *
  * \tparam Kernel Type of the kernel.
  */
@@ -52,7 +56,7 @@ public:
      * \param[in] kernel Kernel.
      */
     explicit kernel_interpolator(const kernel_type& kernel = kernel_type())
-        : kernel_(kernel) {
+        : solver_(kernel) {
         search_kernel_param_auto();
     }
 
@@ -63,7 +67,7 @@ public:
      * \return This.
      */
     auto regularize_with(const value_type& reg_param) -> kernel_interpolator& {
-        interpolator_.regularize_with(reg_param);
+        solver_.regularize_with(reg_param);
         return *this;
     }
 
@@ -73,7 +77,7 @@ public:
      * \return This.
      */
     auto regularize_automatically() -> kernel_interpolator& {
-        interpolator_.regularize_automatically();
+        solver_.regularize_automatically();
         return *this;
     }
 
@@ -83,7 +87,7 @@ public:
      * \return This.
      */
     auto disable_regularization() -> kernel_interpolator& {
-        regularize_with(static_cast<value_type>(0));
+        solver_.regularize_with(static_cast<value_type>(0));
         return *this;
     }
 
@@ -93,7 +97,7 @@ public:
      * \return Regularization parameter.
      */
     [[nodiscard]] auto reg_param() const -> value_type {
-        return interpolator_.reg_param();
+        return solver_.reg_param();
     }
 
     /*!
@@ -104,8 +108,7 @@ public:
      */
     auto fix_kernel_param(const kernel_param_type& kernel_param)
         -> kernel_interpolator& {
-        kernel_.kernel_param(kernel_param);
-        optimizer_.reset();
+        solver_.fix_kernel_param(kernel_param);
         return *this;
     }
 
@@ -115,11 +118,7 @@ public:
      * \return This.
      */
     auto search_kernel_param_auto() -> kernel_interpolator& {
-        if (!optimizer_) {
-            optimizer_ =
-                std::make_unique<impl::kernel_parameter_optimizer<kernel_type>>(
-                    interpolator_, kernel_);
-        }
+        solver_.search_kernel_param_auto();
         return *this;
     }
 
@@ -128,7 +127,9 @@ public:
      *
      * \return Kernel.
      */
-    [[nodiscard]] auto kernel() const -> const kernel_type& { return kernel_; }
+    [[nodiscard]] auto kernel() const -> const kernel_type& {
+        return solver_.kernel();
+    }
 
     /*!
      * \brief Compute internal matrices.
@@ -137,38 +138,12 @@ public:
      * \param[in] data Data vector.
      */
     template <typename Data>
-    void compute(const std::vector<variable_type>& variable_list,
+    void compute(std::vector<variable_type> variable_list,
         const Eigen::MatrixBase<Data>& data) {
-        NUM_COLLECT_ASSERT(
-            static_cast<std::size_t>(data.rows()) == variable_list.size());
-        NUM_COLLECT_ASSERT(data.cols() == 1);
-
-        if (optimizer_) {
-            optimizer_->compute(variable_list, data);
-            kernel_.kernel_param(optimizer_->opt_param());
-        }
-        interpolator_.compute(kernel_, variable_list, data);
-        interpolator_.solve(coeff_);
-
-        variable_list_ = variable_list;
-    }
-
-    /*!
-     * \brief Get the value of the MLE objective function.
-     *
-     * \return Value of the MLE objective function.
-     */
-    [[nodiscard]] auto mle_objective_function_value() const -> value_type {
-        return interpolator_.mle_objective_function_value();
-    }
-
-    /*!
-     * \brief Calculate the coefficient of the kernel common in variables.
-     *
-     * \return Value.
-     */
-    [[nodiscard]] auto common_coeff() const -> value_type {
-        return interpolator_.common_coeff();
+        solver_.compute(variable_list, data);
+        solver_.solve(coeff_);
+        variable_list_ = std::move(variable_list);
+        common_coeff_ = solver_.common_coeff();
     }
 
     /*!
@@ -181,7 +156,7 @@ public:
         -> value_type {
         auto value = static_cast<value_type>(0);
         for (std::size_t i = 0; i < variable_list_.size(); ++i) {
-            value += kernel_(variable, variable_list_[i]) *
+            value += kernel()(variable, variable_list_[i]) *
                 coeff_(static_cast<index_type>(i));
         }
         return value;
@@ -193,13 +168,69 @@ public:
      * \param[in] variable Variable.
      * \return Value.
      */
-    [[nodiscard]] auto operator()(const variable_type& variable) const {
+    [[nodiscard]] auto operator()(const variable_type& variable) const
+        -> value_type {
         return interpolate_on(variable);
     }
 
+    /*!
+     * \brief Evaluate mean and variance for a variable.
+     *
+     * \param[in] variable Variable.
+     * \return Mean and variance.
+     */
+    [[nodiscard]] auto evaluate_mean_and_variance_on(
+        const variable_type& variable) const
+        -> std::pair<value_type, value_type> {
+        Eigen::VectorXd kernel_vec;
+        kernel_vec.resize(static_cast<index_type>(variable_list_.size()));
+        for (std::size_t i = 0; i < variable_list_.size(); ++i) {
+            kernel_vec(static_cast<index_type>(i)) =
+                kernel()(variable, variable_list_[i]);
+        }
+
+        const value_type mean = kernel_vec.dot(coeff_);
+        const value_type variance = common_coeff_ *
+            std::max(kernel()(variable, variable) - calc_reg_term(kernel_vec),
+                static_cast<value_type>(0));
+        return {mean, variance};
+    }
+
+    /*!
+     * \brief Get the value of the MLE objective function
+     * \cite Scheuerer2011.
+     *
+     * \return Value of the MLE objective function.
+     */
+    [[nodiscard]] auto mle_objective_function_value() const -> value_type {
+        return solver_.mle_objective_function_value();
+    }
+
+    /*!
+     * \brief Calculate the coefficient of the kernel common in variables.
+     *
+     * \return Value.
+     */
+    [[nodiscard]] auto common_coeff() const -> value_type {
+        return common_coeff_;
+    }
+
+    /*!
+     * \brief Calculate the regularization term for a vector.
+     *
+     * \tparam InputData Type of the input data.
+     * \param[in] data Data vector.
+     * \return Value.
+     */
+    template <typename InputData>
+    [[nodiscard]] auto calc_reg_term(
+        const Eigen::MatrixBase<InputData>& data) const -> value_type {
+        return solver_.calc_reg_term(data);
+    }
+
 private:
-    //! Kernel.
-    kernel_type kernel_;
+    //! Solver.
+    impl::kernel_coeff_solver<kernel_type> solver_;
 
     //! List of variables.
     std::vector<variable_type> variable_list_;
@@ -207,11 +238,8 @@ private:
     //! Coefficients of kernels for variables.
     Eigen::VectorX<value_type> coeff_{};
 
-    //! Internal interpolator.
-    impl::kernel_interpolator_impl<value_type> interpolator_{};
-
-    //! Optimizer of kernel parameters.
-    std::unique_ptr<impl::kernel_parameter_optimizer<kernel_type>> optimizer_{};
+    //! Coefficient of the kernel common in variables.
+    value_type common_coeff_{};
 };
 
 }  // namespace num_collect::interp::kernel